Move libraries to netstandard and tests to core
