+++ RangeTree.Test/InvalidRangeItemTests.cs
@@ -1,0 +1,17 @@
+﻿namespace RangeTreeTests
+{
+    using System;
+    using RangeTree.Examples;
+    using Xunit;
+
+    public class InvalidRangeItemTests
+    {
+        [Fact]
+        public void ConstructNewRange_ProvideInvalidRange_ExpectArgumentOutOfRangeException()
+        {
+            // Act & Assert
+            Assert.Throws<ArgumentOutOfRangeException>(() => new RangeItem(2, 1));
+            Assert.Throws<ArgumentOutOfRangeException>(() => new RangeItem(2, 1, ""));
+        }
+    }
+}
