+++ RangeTreeExamples/RangeTree.Examples.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard1.3</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\RangeTree\RangeTree.csproj" />
+  </ItemGroup>
+
+</Project>
